requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
reportlab==4.0.7
google-generativeai==0.8.3
python-dotenv==1.0.0
gitpython==3.1.44 